refactor: context provider가 props를 전달받지 않고 직접 데이터를 관리하도록 수정
