+++ src/pages/Layout.tsx
@@ -2,28 +2,12 @@ import styled from '@emotion/styled';
 import { Outlet } from 'react-router-dom';
 import Header from '../components/Header';
 import { CartProvider } from '../context/CartContext';
-import { useState } from 'react';
 import { ErrorMessageProvider } from '../context/ErrorMessageContext';
-import useCartHandler from '../hooks/useCartHandler';
 
 const Layout = () => {
-  const [errorMessage, setErrorMessage] = useState('');
-
-  const handleErrorMessage = (errorMessage: string) => {
-    setErrorMessage(errorMessage);
-  };
-
-  const { cartItemsIds, handleAddCartItemsIds, handleRemoveCartItemsIds } = useCartHandler({
-    handleErrorMessage,
-  });
-
   return (
-    <ErrorMessageProvider errorMessage={errorMessage} handleErrorMessage={handleErrorMessage}>
-      <CartProvider
-        cartItemsIds={cartItemsIds}
-        handleAddCartItemsIds={handleAddCartItemsIds}
-        handleRemoveCartItemsIds={handleRemoveCartItemsIds}
-      >
+    <ErrorMessageProvider>
+      <CartProvider>
         <LayoutContainer>
           <Header />
           <Outlet />
